requirements.txt
grpcio
grpcio-tools
docker
pyyaml